requirements.txt
pandas
openpyxl
psycopg2-binary
python-dotenv
PyYAML
SQLAlchemy